backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
neo4j>=5.15.0
cassandra-driver>=3.28.0
redis>=5.0.0
httpx>=0.25.0
websockets>=12.0
prometheus-fastapi-instrumentator>=6.1.0
structlog>=23.2.0
confluent-kafka>=2.3.0
python-dotenv>=1.0.0
lxml>=4.9.0
jinja2>=3.1.0
